api/pyproject.toml
[tool.poetry]
name = "ebac-djangoframework"
version = "0.1.0"
description = "modulo 4"
authors = ["Mateus H <123660865+MateusHv999@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
django-core = "^1.4.1"
django-filter = "^24.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


